Changed 'long' by 'int' and 'unsigned long' by 'size_t' when necessary. long is 64 bit in GCC when compiling in 64 bits.

+++ include/Step/SimpleTypes.h
@@ -39,7 +39,7 @@
 namespace Step {
 
     //! Integer type
-    typedef long Integer;
+    typedef int Integer;
 
     //! Real type
     typedef double Real;
