+++ src/ifc2x3/SPFWriter.cpp
@@ -757,7 +757,7 @@ bool SPFWriter::write(std::ostream& filestream)
         _callback->setMaximum(vectorOfEntities.size());
     }
     it_vector = vectorOfEntities.begin();
-    long nb=0;
+    size_t nb=0;
     while (it_vector != vectorOfEntities.end()) {
         (*it_vector)->acceptVisitor(this);
         ++it_vector;
